refactor: rename http to service
