+++ src/components/Card/CellCard.tsx
@@ -3,13 +3,13 @@ import styled from 'styled-components'
 import { Link } from 'react-router-dom'
 import AppContext from '../../contexts/App'
 import CopyGreenIcon from '../../assets/copy_green.png'
-import { InputOutput } from '../../http/response/Transaction'
+import { InputOutput } from '../../service/response/Transaction'
 import { startEndEllipsis, hexToUtf8 } from '../../utils/string'
 import { shannonToCkb, copyElementValue } from '../../utils/util'
-import { CellType, fetchScript, fetchCellData } from '../../http/fetcher'
-import { ScriptWrapper } from '../../http/response/Script'
-import { Data } from '../../http/response/Data'
-import { Response } from '../../http/response/Response'
+import { CellType, fetchScript, fetchCellData } from '../../service/fetcher'
+import { ScriptWrapper } from '../../service/response/Script'
+import { Data } from '../../service/response/Data'
+import { Response } from '../../service/response/Response'
 import { localeNumberString } from '../../utils/number'
 import i18n from '../../utils/i18n'
