+++ src/components/Search/index.tsx
@@ -1,10 +1,10 @@
 import React, { useContext, useState, useRef } from 'react'
 import styled, { css } from 'styled-components'
 import AppContext from '../../contexts/App'
-import { fetchSearchResult } from '../../http/fetcher'
-import { BlockWrapper } from '../../http/response/Block'
-import { TransactionWrapper } from '../../http/response/Transaction'
-import { AddressWrapper } from '../../http/response/Address'
+import { fetchSearchResult } from '../../service/fetcher'
+import { BlockWrapper } from '../../service/response/Block'
+import { TransactionWrapper } from '../../service/response/Transaction'
+import { AddressWrapper } from '../../service/response/Address'
 import browserHistory from '../../routes/history'
 import SearchLogo from '../../assets/search.png'
 import { searchTextCorrection } from '../../utils/string'
