+++ src/components/Transaction/TransactionCellList/index.tsx
@@ -1,7 +1,7 @@
 import React, { ReactNode } from 'react'
 import { Link } from 'react-router-dom'
 import TransactionCellListPanel from './styled'
-import { Transaction, InputOutput } from '../../../http/response/Transaction'
+import { Transaction, InputOutput } from '../../../service/response/Transaction'
 
 interface TransactionCellListProps {
   cells: InputOutput[]
