+++ src/pages/Address/index.tsx
@@ -29,15 +29,15 @@ import TransactionsIcon from '../../assets/transactions_green.png'
 import ItemPointIcon from '../../assets/item_point.png'
 import AddressHashIcon from '../../assets/lock_hash_address.png'
 import BlockPendingRewardIcon from '../../assets/block_pending_reward.png'
-import { Address, AddressWrapper } from '../../http/response/Address'
-import { Script } from '../../http/response/Script'
-import { Response } from '../../http/response/Response'
-import { TransactionWrapper } from '../../http/response/Transaction'
-import { fetchAddressInfo, fetchTransactionsByAddress, fetchTipBlockNumber } from '../../http/fetcher'
+import { Address, AddressWrapper } from '../../service/response/Address'
+import { Script } from '../../service/response/Script'
+import { Response } from '../../service/response/Response'
+import { TransactionWrapper } from '../../service/response/Transaction'
+import { fetchAddressInfo, fetchTransactionsByAddress, fetchTipBlockNumber } from '../../service/fetcher'
 import { copyElementValue, shannonToCkb } from '../../utils/util'
 import { parsePageNumber, startEndEllipsis } from '../../utils/string'
 import TransactionCard from '../../components/Transaction/TransactionCard/index'
-import { StatisticsWrapper } from '../../http/response/Statistics'
+import { StatisticsWrapper } from '../../service/response/Statistics'
 import { localeNumberString } from '../../utils/number'
 import i18n from '../../utils/i18n'
 import { isMobile } from '../../utils/screen'
