+++ src/pages/BlockDetail/index.tsx
@@ -42,11 +42,11 @@ import NextBlockGreyIcon from '../../assets/right_arrow_grey.png'
 import MouseIcon from '../../assets/block_mouse.png'
 import TransactionsRootIcon from '../../assets/transactions_root.png'
 import WitnessRootIcon from '../../assets/witness_root.png'
-import { Block, BlockWrapper, RewardStatus, TransactionFeeStatus } from '../../http/response/Block'
+import { Block, BlockWrapper, RewardStatus, TransactionFeeStatus } from '../../service/response/Block'
 import { parseSimpleDate } from '../../utils/date'
-import { Response } from '../../http/response/Response'
-import { TransactionWrapper } from '../../http/response/Transaction'
-import { fetchBlock, fetchTransactionsByBlockHash } from '../../http/fetcher'
+import { Response } from '../../service/response/Response'
+import { TransactionWrapper } from '../../service/response/Transaction'
+import { fetchBlock, fetchTransactionsByBlockHash } from '../../service/fetcher'
 import { copyElementValue, shannonToCkb } from '../../utils/util'
 import { startEndEllipsis, parsePageNumber } from '../../utils/string'
 import browserHistory from '../../routes/history'
