+++ src/pages/BlockList/index.tsx
@@ -21,9 +21,9 @@ import TransactionIcon from '../../assets/transactions.png'
 import BlockRewardIcon from '../../assets/block_reward_white.png'
 import MinerIcon from '../../assets/miner.png'
 import TimestampIcon from '../../assets/timestamp.png'
-import { fetchBlockList } from '../../http/fetcher'
-import { BlockWrapper } from '../../http/response/Block'
-import { Response } from '../../http/response/Response'
+import { fetchBlockList } from '../../service/fetcher'
+import { BlockWrapper } from '../../service/response/Block'
+import { Response } from '../../service/response/Response'
 import { shannonToCkb } from '../../utils/util'
 import { parsePageNumber } from '../../utils/string'
 import { CachedKeys } from '../../utils/const'
