+++ src/pages/Home/index.tsx
@@ -18,10 +18,10 @@ import MinerIcon from '../../assets/miner.png'
 import TimestampIcon from '../../assets/timestamp.png'
 import MoreLeftIcon from '../../assets/more_left.png'
 import MoreRightIcon from '../../assets/more_right.png'
-import { fetchBlocks, fetchStatistics } from '../../http/fetcher'
-import { BlockWrapper } from '../../http/response/Block'
-import { StatisticsWrapper, Statistics } from '../../http/response/Statistics'
-import { Response } from '../../http/response/Response'
+import { fetchBlocks, fetchStatistics } from '../../service/fetcher'
+import { BlockWrapper } from '../../service/response/Block'
+import { StatisticsWrapper, Statistics } from '../../service/response/Statistics'
+import { Response } from '../../service/response/Response'
 import { shannonToCkb } from '../../utils/util'
 import { parseTime, parseSimpleDate } from '../../utils/date'
 import { BLOCK_POLLING_TIME, CachedKeys } from '../../utils/const'
