+++ src/pages/Transaction/Script.tsx
@@ -2,15 +2,15 @@ import React, { useContext, useReducer } from 'react'
 import { Link } from 'react-router-dom'
 import AppContext from '../../contexts/App'
 import CopyGreenIcon from '../../assets/copy_green.png'
-import { Response } from '../../http/response/Response'
-import { ScriptWrapper } from '../../http/response/Script'
-import { CellType, fetchScript, fetchCellData } from '../../http/fetcher'
+import { Response } from '../../service/response/Response'
+import { ScriptWrapper } from '../../service/response/Script'
+import { CellType, fetchScript, fetchCellData } from '../../service/fetcher'
 import { shannonToCkb, copyElementValue } from '../../utils/util'
 import { hexToUtf8, parseLongAddressHash } from '../../utils/string'
 import { localeNumberString } from '../../utils/number'
 import i18n from '../../utils/i18n'
-import { InputOutput } from '../../http/response/Transaction'
-import { Data } from '../../http/response/Data'
+import { InputOutput } from '../../service/response/Transaction'
+import { Data } from '../../service/response/Data'
 
 enum CellState {
   NONE,
