+++ src/pages/Transaction/index.tsx
@@ -21,13 +21,13 @@ import TransactionIcon from '../../assets/transaction_fee.png'
 import CopyIcon from '../../assets/copy.png'
 import StatusIcon from '../../assets/transcation_status.png'
 import { parseSimpleDate } from '../../utils/date'
-import { Response } from '../../http/response/Response'
-import { Transaction, InputOutput, TransactionWrapper } from '../../http/response/Transaction'
-import { CellType, fetchTransactionByHash, fetchTipBlockNumber } from '../../http/fetcher'
+import { Response } from '../../service/response/Response'
+import { Transaction, InputOutput, TransactionWrapper } from '../../service/response/Transaction'
+import { CellType, fetchTransactionByHash, fetchTipBlockNumber } from '../../service/fetcher'
 import { copyElementValue, formatConfirmation, shannonToCkb } from '../../utils/util'
 import CellCard from '../../components/Card/CellCard'
 import ScriptComponent from './Script'
-import { StatisticsWrapper } from '../../http/response/Statistics'
+import { StatisticsWrapper } from '../../service/response/Statistics'
 import { localeNumberString } from '../../utils/number'
 import { isMobile } from '../../utils/screen'
